requirements.txt
pynput
opencv-python
easyocr
pyautogui
pywebview
torch
numpy==1.22.4